Add auto-generated component schemas in openapi.yaml from models
Scan for all models in application/models, and generate a component
schema for each. Some of them were internals or not useful otherwise
(e.g. no table match in the DB), so delete them. Also, comment out
some maybe-not-useful ones, but not deleting them because they may
be used in the future, so to keep it in the commit history for
faster access.

Add the PHP schema-generator also. Note that, as the code is ugly,
it may be removed in the next commit, for keeping it in the history
only.