Add support for question resources in openapi.yaml
The path of a question is is /surveys/{survey_id}/questions
/{question_id}. It seems to be an independent concept from a question
group, so it is not a sub-path of a group. You can see this in the
LimeSurvey UI itself: When you want to access questions list, you
do not have to go through a specific group, although you can.