Handle all paths starting with /restApi
Subscribe to beforeControllerAction event, and if the base path
(i.e. the path after the script url) starts with /restApi, catch it
and handle it ourselves.

Handling it is by using a new Router class. The class itself uses
the Symfony routing component, and uses a new Routes class holding
the routing information. Router::route() is the main function to
handle routing, and returns the handler controller and method name.
Then Plugin calls it and prints its results.

The API path is something like '/restApi/v0/login/bearer_token'.