Improve array dual scale answers validation
Convert AnswerValidatorBuilder from a static class to non-static.
The reason is, to know what should be done with questions with soft
mandatory, we need to pass in - even currently is unsued - a param
to specify whether to suppose them as mandatory or not (e.g. to get
user confirmation later); which is where a constructor comes into
play.