Refactor the generators (i.e. data converter from and to a DB record)
Add support for all question types, except for uploading a file as
an answer. Group method to question type mapping to look easier on
the eyes, and better to understand and deduplicate functionality.

Extend the line limit from previous 100 to 120, as 100 is limiting
for long function prototypes, specially with long names.